+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-08-31 12:02:53
-// Commit: feat(forms): add async forms stream provider
+// Updated at: 2026-08-31 12:38:25
+// Commit: feat(forms): integrate custom telemetry logger in forms
 
 class formsComponentState {
-  final String commitRef = '4068';
-  final String lastMessage = 'feat(forms): add async forms stream provider';
-  final DateTime timestamp = DateTime.parse('2026-08-31T12:02:53+05:00');
+  final String commitRef = '4070';
+  final String lastMessage = 'feat(forms): integrate custom telemetry logger in forms';
+  final DateTime timestamp = DateTime.parse('2026-08-31T12:38:25+05:00');
 }